requirements.txt
httpx
pillow
coco-lib
pandas
Pillow
torch
torchvision
tqdm
transformers
pybioclip
